+++ src/AIPrompt.Core/Interfaces/IDatabaseInitializerService.cs
@@ -1,0 +1,6 @@
+namespace AIPrompt.Core.Interfaces;
+
+public interface IDatabaseInitializerService
+{
+    Task InitializeAsync(CancellationToken cancellationToken = default);
+}